Parser v1
Note: This only works for the messages.json file.
Place the messages.json and the py file in the same directory and run the script. The script prompts you to enter your username just for verification and then you can select the conversation you want to extract. The conversations get extracted as an SQLite file in the same directory which can be opened using DB browser for SQLite or converted to XLS and CSV.